+++ cmd/network_start.go
@@ -4,10 +4,17 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"time"
 
 	"code.vegaprotocol.io/vegacapsule/nomad"
 	"code.vegaprotocol.io/vegacapsule/state"
 	"github.com/spf13/cobra"
+)
+
+const defaultTimeout = 300 * time.Second
+
+var (
+	timeout time.Duration
 )
 
 var netStartCmd = &cobra.Command{
@@ -23,13 +30,25 @@ var netStartCmd = &cobra.Command{
 			return networkNotBootstrappedErr("start")
 		}
 
-		updatedNetState, err := netStart(context.Background(), *netState)
+		ctx, cancel := context.WithTimeout(context.Background(), timeout)
+		defer cancel()
+		updatedNetState, err := netStart(ctx, *netState)
+		// We want state saved even if the network is not started properly
+		defer saveNetworkState(updatedNetState)
 		if err != nil {
 			return fmt.Errorf("failed to start network: %w", err)
 		}
 
-		return updatedNetState.Persist()
+		return err
 	},
+}
+
+func init() {
+	netStartCmd.PersistentFlags().DurationVar(&timeout,
+		"timeout",
+		defaultTimeout,
+		"Network start timeout",
+	)
 }
 
 func netStart(ctx context.Context, state state.NetworkState) (*state.NetworkState, error) {
@@ -43,10 +62,14 @@ func netStart(ctx context.Context, state state.NetworkState) (*state.NetworkStat
 	nomadRunner := nomad.NewJobRunner(nomadClient)
 
 	res, err := nomadRunner.StartNetwork(ctx, state.Config, state.GeneratedServices)
+	// if network state returned, save it in current state regardless of an error
+	if res != nil {
+		state.RunningJobs = res
+	}
+
 	if err != nil {
 		return nil, fmt.Errorf("failed to start nomad network: %s", err)
 	}
-	state.RunningJobs = res
 
 	log.Println("starting network success")
 	return &state, nil