+++ cmd/nodes_reset_state.go
@@ -1,14 +1,19 @@
 package cmd
 
 import (
+	"context"
 	"fmt"
 	"io"
+	"log"
 	"os"
 
 	"code.vegaprotocol.io/vegacapsule/commands"
+	"code.vegaprotocol.io/vegacapsule/nomad"
 	"code.vegaprotocol.io/vegacapsule/state"
 	"github.com/spf13/cobra"
 )
+
+var remoteResetState bool
 
 var nodesUnsafeResetAllCmd = &cobra.Command{
 	Use:   "unsafe-reset-all",
@@ -23,18 +28,42 @@ var nodesUnsafeResetAllCmd = &cobra.Command{
 			return networkNotBootstrappedErr("state unsafe-reset-all")
 		}
 
-		r, err := commands.ResetNodeSetsData(
-			*netState.Config.VegaBinary,
-			netState.GeneratedServices.NodeSets.ToSlice(),
-		)
+		var cmdResult io.Reader
+
+		if !remoteResetState {
+			log.Print("Running unsafe-reset-all on local machine")
+			cmdResult, err = commands.ResetNodeSetsData(
+				*netState.Config.VegaBinary,
+				netState.GeneratedServices.NodeSets.ToSlice(),
+			)
+		} else {
+			log.Print("Running unsafe-reset-all on remote cluster")
+			var nomadClient *nomad.Client
+			nomadClient, err = nomad.NewClient(nil)
+			if err != nil {
+				return fmt.Errorf("failed to create nomad client: %w", err)
+			}
+
+			commandRunner := nomad.NewCommandRunner(nomadClient)
+			cmdResult, err = commandRunner.NetworkUnsafeResetAll(context.Background(), netState.GeneratedServices.NodeSets.ToSlice())
+		}
+
 		if err != nil {
 			return fmt.Errorf("failed to reset node sets: %w", err)
 		}
 
-		if _, err := io.Copy(os.Stdout, r); err != nil {
+		if _, err := io.Copy(os.Stdout, cmdResult); err != nil {
 			return fmt.Errorf("failed to write command output: %w", err)
 		}
 
 		return nil
 	},
+}
+
+func init() {
+	nodesUnsafeResetAllCmd.PersistentFlags().BoolVar(&remoteResetState,
+		"remote",
+		false,
+		"Determines, whether the command should be executed locally or remotely",
+	)
 }